refactor(httpEnvelopingRequest): export
