+++ src/definitions/httpEnvelopingRequest.type.ts
