+++ src/definitions/index.ts
@@ -2,3 +2,4 @@ export * from './errorMessages.const';
 export * from './event.type';
 export * from './server.const';
 export * from './token.type';
+export * from './httpEnvelopingRequest.type';
